+++ ts/features/walletV3/payment/screens/WalletPaymentConfirmScreen.tsx
@@ -122,7 +122,6 @@ const WalletPaymentConfirmScreen = () => {
         O.chainNullableK(method => method.details)
       ),
       selectedPsp: selectedPspOption,
-      selectedMethod: selectedMethodOption,
       paymentDetails: pipe(paymentDetailsPot, pot.toOption)
     }),
     O.fold(